Put config in orchestrator
